vdjapi/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-auth==0.9.1
djangorestframework==3.6.4
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.10.1
djangorestframework-jwt==1.11.0
idna==2.6
PyJWT==1.5.2
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.4
rest-condition==1.0.3
six==1.10.0
urllib3==1.22
